Rename Contents to CharSequenceContent
